text-to-speech/requirements.txt
python-dotenv
openai
fastapi
uvicorn[standard]